LPS-134118 Move the useObserveRect hook to its own file
